changed package name to comprl
